feat: unittest 추가&refactor
